+++ src/components/TodoList.vue
@@ -1,7 +1,7 @@
 <template>
   <section>
       <ul>
-        <li v-for="(todoItem, index) in todoItems" class="shadow">
+        <li v-for="(todoItem, index) in propsdata" class="shadow">
           <i class="checkBtn fa fa-check" aria-hidden="true"></i>
           {{ todoItem }}
           <span class="removeBtn" type="button" @click="removeTodo(todoItem, index)">
@@ -14,22 +14,11 @@
 
 <script>
     export default {
-        data() {
-            return {
-               todoItems: []
-            }
-        },
-        created() {
-            if(localStorage.length > 0){
-                for(var i=0; i<localStorage.length; i++){
-                    this.todoItems.push(localStorage.key(i))
-                }
-            }
-        },
+        props : ['propsdata'],
         methods: {
             removeTodo(todoItem, index) {
                 localStorage.removeItem(todoItem);
-                this.todoItems.splice(index, 1); // 배열의 특정 인덱스 삭제
+                this.propsdata.splice(index, 1); // 배열의 특정 인덱스 삭제
             }
         }
     }